# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Akure Airport Staff Cooperative Society API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile (member record for current clerk user)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
        "404":
          description: Not found
  /auth/register:
    post:
      operationId: registerMember
      tags: [auth]
      summary: Self-register a new member (locked until admin activates)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Bad request

  # ── MEMBERS ──────────────────────────────────────────────────────────────────
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List all loans (Admin+)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, admin_approved, auditor_approved, super_admin_approved, disbursed, rejected]
        - in: query
          name: memberId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Apply for a loan (Member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanBody"
      responses:
        "201":
          description: Loan application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
...
  /loans/calculate:
    post:
      operationId: calculateLoan
      tags: [loans]
      summary: Calculate loan repayment preview (any authenticated user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanCalculateBody"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanCalculation"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get a single loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
  /loans/{id}/approve:
    post:
      operationId: approveLoan
      tags: [loans]
      summary: Approve a loan (role-dependent stage)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanActionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/reject:
    post:
      operationId: rejectLoan
      tags: [loans]
      summary: Reject a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanActionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/disburse:
    post:
      operationId: disburseLoan
      tags: [loans]
      summary: Mark loan as disbursed (Treasurer)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanActionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/repayments:
    get:
      operationId: getLoanRepayments
      tags: [loans]
      summary: Get repayment history for a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # ── STORE ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        staffId:
          type: ["string", "null"]
        role:
          type: string
          enum: [member, admin, financial_auditor, treasurer, super_admin]
        status:
          type: string
          enum: [pending, active, inactive]
        savingsBalance:
          type: number
        totalLoanBalance:
          type: number
        totalStoreDebt:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - status
        - savingsBalance
        - totalLoanBalance
        - totalStoreDebt
        - createdAt
        - updatedAt
    MemberProfile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        staffId:
          type: ["string", "null"]
        role:
          type: string
          enum: [member, admin, financial_auditor, treasurer, super_admin]
        status:
          type: string
          enum: [pending, active, inactive]
        savingsBalance:
          type: number
        totalLoanBalance:
          type: number
        totalStoreDebt:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - fullName
        - email
        - role
        - status
        - savingsBalance
        - totalLoanBalance
        - totalStoreDebt
        - createdAt
    RegisterMemberBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        staffId:
          type: string
      required: [fullName]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        type:
          type: string
          enum: [savings, loan_repayment, store_repayment]
        amount:
          type: number
        description:
          type: ["string", "null"]
        uploadRecordId:
          type: ["integer", "null"]
        month:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - type
        - amount
        - createdAt

    # ── EXCEL UPLOAD ─────────────────────────────────────────────────────────────
...
    Loan:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        amount:
          type: number
        interestRate:
          type: number
        interestAmount:
          type: number
        totalRepayable:
          type: number
        monthlyRepayment:
          type: number
        tenureMonths:
          type: integer
        purpose:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, admin_approved, auditor_approved, super_admin_approved, disbursed, rejected]
        outstandingBalance:
          type: number
        adminApprovedAt:
          type: ["string", "null"]
          format: date-time
        adminApprovedBy:
          type: ["integer", "null"]
        auditorApprovedAt:
          type: ["string", "null"]
          format: date-time
        auditorApprovedBy:
          type: ["integer", "null"]
        superAdminApprovedAt:
          type: ["string", "null"]
          format: date-time
        superAdminApprovedBy:
          type: ["integer", "null"]
        disbursedAt:
          type: ["string", "null"]
          format: date-time
        disbursedBy:
          type: ["integer", "null"]
        rejectedAt:
          type: ["string", "null"]
          format: date-time
        rejectedBy:
          type: ["integer", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - amount
        - interestRate
        - interestAmount
        - totalRepayable
        - monthlyRepayment
        - tenureMonths
        - status
        - outstandingBalance
        - createdAt
        - updatedAt
    CreateLoanBody:
      type: object
      properties:
        amount:
          type: number
        tenureMonths:
          type: integer
        purpose:
          type: string
      required: [amount, tenureMonths]
    LoanActionBody:
      type: object
      properties:
        notes:
          type: string
    LoanCalculateBody:
      type: object
      properties:
        amount:
          type: number
        tenureMonths:
          type: integer
      required: [amount, tenureMonths]
    LoanCalculation:
      type: object
      properties:
        principal:
          type: number
        interestRate:
          type: number
        interestAmount:
          type: number
        totalRepayable:
          type: number
        monthlyRepayment:
          type: number
        tenureMonths:
          type: integer
      required:
        - principal
        - interestRate
        - interestAmount
        - totalRepayable
        - monthlyRepayment
        - tenureMonths

    # ── STORE ────────────────────────────────────────────────────────────────────
...